# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CoopData API",
    "description": "Cooperative Data Management Platform API with RBAC hierarchy (Ministry \u2192 Federation \u2192 Apex \u2192 Cooperative)",
    "contact": {
      "name": "CoopData Team",
      "email": "support@coopdata.org"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/apex/coop-profiles/{id}": {
      "get": {
        "tags": [
          "Cooperatives"
        ],
        "operationId": "get_cooperative_profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cooperative ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cooperative profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CooperativeProfileResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Cooperative not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Cooperatives"
        ],
        "operationId": "delete_cooperative_profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cooperative ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cooperative deleted"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Cooperative not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Cooperatives"
        ],
        "operationId": "update_cooperative_profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cooperative ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCooperativeProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Cooperative updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CooperativeProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Cooperative not found"
          },
          "409": {
            "description": "reg_no conflict"
          }
        }
      }
    },
...
    "/api/v1/apex/cooperatives/{id}/members": {
      "get": {
        "tags": [
          "Apex"
        ],
        "operationId": "list_cooperative_members",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cooperative (Subgroup) ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of cooperative members",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberResponse"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Apex"
        ],
        "operationId": "add_cooperative_member",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cooperative (Subgroup) ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Member added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or wrong role"
          },
          "403": {
            "description": "Forbidden - apex role required or scope violation"
          },
          "409": {
            "description": "User already in a group"
          }
        }
      }
    },
...
    "/api/v1/cooperative/extraction-jobs/{id}": {
      "get": {
        "tags": [
          "Cooperative"
        ],
        "operationId": "get_extraction_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Extraction job ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Extraction job status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionJobResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/v1/cooperative/financial-statement/upload": {
      "post": {
        "tags": [
          "Cooperative"
        ],
        "operationId": "upload_financial_statement",
        "responses": {
          "202": {
            "description": "Upload accepted, extraction queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or unsupported file type"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Cooperative profile not found"
          },
          "409": {
            "description": "Submission already exists for this year"
          }
        }
      }
    },
    "/api/v1/cooperative/financial-statements/{id}": {
      "get": {
        "tags": [
          "Cooperative"
        ],
        "operationId": "get_financial_statement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Financial statement ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Financial statement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialStatementResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/v1/cooperative/financial-statements/{id}/line-items": {
      "get": {
        "tags": [
          "Cooperative"
        ],
        "operationId": "list_line_items",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Financial statement ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Line items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LineItemResponse"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Cooperative"
        ],
        "operationId": "update_line_items",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Financial statement ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineItemBulkUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Line items updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LineItemResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/cooperative/non-financial/farm-coop": {
      "get": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "list_farm_coop",
        "parameters": [
          {
            "name": "submission_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List farm coop records",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedFarmCoopResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "create_farm_coop",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFarmCoopRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Farm coop record created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmCoopResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cooperative/non-financial/farm-coop/{id}": {
      "get": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "get_farm_coop",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Farm coop record ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Farm coop record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmCoopResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "update_farm_coop",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Farm coop record ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFarmCoopRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Farm coop record updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmCoopResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "delete_farm_coop",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Farm coop record ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Farm coop record deleted"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cooperative/non-financial/fixed-deposits": {
      "get": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "list_fixed_deposits",
        "parameters": [
          {
            "name": "submission_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List fixed deposits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedFixedDepositsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "create_fixed_deposit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFixedDepositRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Fixed deposit created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixedDepositResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cooperative/non-financial/fixed-deposits/{id}": {
      "get": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "get_fixed_deposit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Fixed Deposit ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fixed deposit found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixedDepositResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "update_fixed_deposit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Fixed Deposit ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFixedDepositRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Fixed deposit updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixedDepositResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "delete_fixed_deposit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Fixed Deposit ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Fixed deposit deleted"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cooperative/non-financial/loans": {
      "get": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "list_loans",
        "parameters": [
          {
            "name": "submission_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List loans",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedLoansResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "create_loan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Loan created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cooperative/non-financial/loans/{id}": {
      "get": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "get_loan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Loan ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Loan found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "update_loan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Loan ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLoanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Loan updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "delete_loan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Loan ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Loan deleted"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cooperative/non-financial/upload": {
      "post": {
        "tags": [
          "Non-Financial"
        ],
        "operationId": "upload_non_financial",
        "responses": {
          "201": {
            "description": "Upload processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NfUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing file, submission_id, or no valid sheets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cooperative/stats": {
      "get": {
        "operationId": "get_cooperative_stats",
        "tags": [
          "Cooperative"
        ],
        "responses": {
          "200": {
            "description": "Cooperative dashboard stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CooperativeStatsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AddMemberRequest": {
        "type": "object",
        "required": [
          "email",
          "first_name",
          "last_name",
          "role"
        ],
        "properties": {
          "assigned_dimensions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "email": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
      "MemberResponse": {
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "first_name": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "last_name": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "username": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}